refactor: move composer to utils
