+++ src/flatprot/utils/__init__.py
@@ -1,0 +1,2 @@
+# Copyright 2025 Tobias Olenyi.
+# SPDX-License-Identifier: Apache-2.0
