+++ src/flatprot/utils/structure.py
